+++ integration/cache_test.go
@@ -23,8 +23,6 @@ import (
 	"k8s.io/apimachinery/pkg/util/wait"
 
 	"github.com/GoogleContainerTools/skaffold/integration/skaffold"
-	"github.com/GoogleContainerTools/skaffold/proto"
-	"github.com/GoogleContainerTools/skaffold/testutil"
 )
 
 func TestCacheAPITriggers(t *testing.T) {
@@ -39,33 +37,16 @@ func TestCacheAPITriggers(t *testing.T) {
 	defer deleteNs()
 
 	rpcAddr := randomPort()
-
 	stop := skaffold.Dev("--rpc-port", rpcAddr).InDir("examples/getting-started").InNs(ns.Name).RunBackground(t)
 	defer stop()
 
-	client, shutdown := setupRPCClient(t, rpcAddr)
+	// Ensure we see a build triggered in the event log
+	_, entries, shutdown := apiEvents(t, rpcAddr)
 	defer shutdown()
 
-	stream, err := readEventAPIStream(client, t, readRetries)
-	if stream == nil {
-		t.Fatalf("error retrieving event log: %v\n", err)
-	}
-
-	// read entries from the log
-	entries := make(chan *proto.LogEntry)
-	go func() {
-		for {
-			entry, _ := stream.Recv()
-			if entry != nil {
-				entries <- entry
-			}
-		}
-	}()
-
-	// Ensure we see a build triggered in the event log
-	err = wait.PollImmediate(time.Millisecond*500, 2*time.Minute, func() (bool, error) {
+	err := wait.PollImmediate(time.Millisecond*500, 2*time.Minute, func() (bool, error) {
 		e := <-entries
 		return e.GetEvent().GetBuildEvent().GetArtifact() == "gcr.io/k8s-skaffold/skaffold-example", nil
 	})
-	testutil.CheckError(t, false, err)
+	failNowIfError(t, err)
 }
