+++ integration/dev_test.go
@@ -93,59 +93,42 @@ func TestDevAPITriggers(t *testing.T) {
 	// Run skaffold build first to fail quickly on a build failure
 	skaffold.Build().InDir("testdata/dev").RunOrFail(t)
 
-	ns, k8sClient, deleteNs := SetupNamespace(t)
+	ns, client, deleteNs := SetupNamespace(t)
 	defer deleteNs()
 
 	rpcAddr := randomPort()
-
 	stop := skaffold.Dev("--auto-build=false", "--auto-sync=false", "--auto-deploy=false", "--rpc-port", rpcAddr, "--cache-artifacts=false").InDir("testdata/dev").InNs(ns.Name).RunBackground(t)
 	defer stop()
 
-	client, shutdown := setupRPCClient(t, rpcAddr)
+	rpcClient, entries, shutdown := apiEvents(t, rpcAddr)
 	defer shutdown()
-
-	stream, err := readEventAPIStream(client, t, readRetries)
-	if stream == nil {
-		t.Fatalf("error retrieving event log: %v\n", err)
-	}
 
 	// throw away first 5 entries of log (from first run of dev loop)
 	for i := 0; i < 5; i++ {
-		stream.Recv()
+		<-entries
 	}
 
-	// read entries from the log
-	entries := make(chan *proto.LogEntry)
-	go func() {
-		for {
-			entry, _ := stream.Recv()
-			if entry != nil {
-				entries <- entry
-			}
-		}
-	}()
-
-	dep := k8sClient.GetDeployment("test-dev")
+	dep := client.GetDeployment("test-dev")
 
 	// Make a change to foo
 	Run(t, "testdata/dev", "sh", "-c", "echo bar > foo")
 
 	// Issue a build trigger
-	client.Execute(context.Background(), &proto.UserIntentRequest{
+	rpcClient.Execute(context.Background(), &proto.UserIntentRequest{
 		Intent: &proto.Intent{
 			Build: true,
 		},
 	})
 
 	// Ensure we see a build triggered in the event log
-	err = wait.PollImmediate(time.Millisecond*500, 2*time.Minute, func() (bool, error) {
+	err := wait.PollImmediate(time.Millisecond*500, 2*time.Minute, func() (bool, error) {
 		e := <-entries
 		return e.GetEvent().GetBuildEvent().GetArtifact() == "gcr.io/k8s-skaffold/test-dev", nil
 	})
-	testutil.CheckError(t, false, err)
+	failNowIfError(t, err)
 
 	// Issue a deploy trigger
-	client.Execute(context.Background(), &proto.UserIntentRequest{
+	rpcClient.Execute(context.Background(), &proto.UserIntentRequest{
 		Intent: &proto.Intent{
 			Deploy: true,
 		},
@@ -156,14 +139,14 @@ func TestDevAPITriggers(t *testing.T) {
 		e := <-entries
 		return e.GetEvent().GetDeployEvent().GetStatus() == "In Progress", nil
 	})
-	testutil.CheckError(t, false, err)
+	failNowIfError(t, err)
 
 	// Make sure the old Deployment and the new Deployment are different
 	err = wait.PollImmediate(time.Millisecond*500, 10*time.Minute, func() (bool, error) {
-		newDep := k8sClient.GetDeployment("test-dev")
+		newDep := client.GetDeployment("test-dev")
 		return dep.GetGeneration() != newDep.GetGeneration(), nil
 	})
-	testutil.CheckError(t, false, err)
+	failNowIfError(t, err)
 }
 
 func TestDevPortForward(t *testing.T) {
@@ -178,47 +161,23 @@ func TestDevPortForward(t *testing.T) {
 	defer deleteNs()
 
 	rpcAddr := randomPort()
-	env := []string{fmt.Sprintf("TEST_NS=%s", ns.Name)}
-	cmd := skaffold.Dev("--port-forward", "--rpc-port", rpcAddr).InDir("examples/microservices").InNs(ns.Name).WithEnv(env)
-	stop := cmd.RunBackground(t)
+	stop := skaffold.Dev("--port-forward", "--rpc-port", rpcAddr).InDir("examples/microservices").InNs(ns.Name).RunBackground(t)
 	defer stop()
 
-	client, shutdown := setupRPCClient(t, rpcAddr)
+	_, entries, shutdown := apiEvents(t, rpcAddr)
 	defer shutdown()
 
-	// create a grpc connection. Increase number of reties for helm.
-	stream, err := readEventAPIStream(client, t, 20)
-	if stream == nil {
-		t.Fatalf("error retrieving event log: %v\n", err)
-	}
+	waitForPortForwardEvent(t, entries, "leeroy-app", "service", ns.Name, "leeroooooy app!!\n")
 
-	// read entries from the log
-	entries := make(chan *proto.LogEntry)
-	go func() {
-		for {
-			entry, _ := stream.Recv()
-			if entry != nil {
-				entries <- entry
-			}
-		}
-	}()
-
-	originalResponse := "leeroooooy app!!"
-	replacementResponse := "test string"
-
-	waitForPortForwardEvent(t, entries, "leeroy-app", "service", ns.Name, originalResponse+"\n")
-
-	original, perms, fErr := replaceInFile(originalResponse, replacementResponse, "examples/microservices/leeroy-app/app.go")
-	if fErr != nil {
-		t.Error(fErr)
-	}
+	original, perms, fErr := replaceInFile("leeroooooy app!!", "test string", "examples/microservices/leeroy-app/app.go")
+	failNowIfError(t, fErr)
 	defer func() {
 		if original != nil {
 			ioutil.WriteFile("examples/microservices/leeroy-app/app.go", original, perms)
 		}
 	}()
 
-	waitForPortForwardEvent(t, entries, "leeroy-app", "service", ns.Name, replacementResponse+"\n")
+	waitForPortForwardEvent(t, entries, "leeroy-app", "service", ns.Name, "test string\n")
 }
 
 func TestDevPortForwardGKELoadBalancer(t *testing.T) {
@@ -234,29 +193,11 @@ func TestDevPortForwardGKELoadBalancer(t *testing.T) {
 
 	rpcAddr := randomPort()
 	env := []string{fmt.Sprintf("TEST_NS=%s", ns.Name)}
-	cmd := skaffold.Dev("--port-forward", "--rpc-port", rpcAddr).InDir("testdata/gke_loadbalancer").InNs(ns.Name).WithEnv(env)
-	stop := cmd.RunBackground(t)
+	stop := skaffold.Dev("--port-forward", "--rpc-port", rpcAddr).InDir("testdata/gke_loadbalancer").InNs(ns.Name).WithEnv(env).RunBackground(t)
 	defer stop()
 
-	client, shutdown := setupRPCClient(t, rpcAddr)
+	_, entries, shutdown := apiEvents(t, rpcAddr)
 	defer shutdown()
-
-	// create a grpc connection. Increase number of reties for helm.
-	stream, err := readEventAPIStream(client, t, 20)
-	if stream == nil {
-		t.Fatalf("error retrieving event log: %v\n", err)
-	}
-
-	// read entries from the log
-	entries := make(chan *proto.LogEntry)
-	go func() {
-		for {
-			entry, _ := stream.Recv()
-			if entry != nil {
-				entries <- entry
-			}
-		}
-	}()
 
 	waitForPortForwardEvent(t, entries, "gke-loadbalancer", "service", ns.Name, "hello!!\n")
 }
@@ -338,22 +279,6 @@ func replaceInFile(target, replacement, filepath string) ([]byte, os.FileMode, e
 	err = ioutil.WriteFile(filepath, []byte(newContents), 0)
 
 	return original, fInfo.Mode(), err
-}
-
-func readEventAPIStream(client proto.SkaffoldServiceClient, t *testing.T, retries int) (proto.SkaffoldService_EventLogClient, error) {
-	t.Helper()
-	// read the event log stream from the skaffold grpc server
-	var stream proto.SkaffoldService_EventLogClient
-	var err error
-	for i := 0; i < retries; i++ {
-		stream, err = client.EventLog(context.Background())
-		if err != nil {
-			t.Logf("waiting for connection...")
-			time.Sleep(waitTime)
-			continue
-		}
-	}
-	return stream, err
 }
 
 func TestDev_WithKubecontextOverride(t *testing.T) {
